requirements.txt
Flask==2.0.2
Werkzeug==2.2.3
openai-whisper
torch==1.10.1
numpy<2.0.0
